change heatmap coords migration

+++ db/migrate/20110208232847_create_heatmap_coords.rb
@@ -1,9 +1,9 @@
 class CreateHeatmapCoords < ActiveRecord::Migration
   def self.up
     create_table :heatmap_coords do |t|
-      t.column :coord_x,       :double
-      t.column :coord_y,       :double
-      t.column :coord_radio,   :double
+      t.column :coord_x,       :decimal
+      t.column :coord_y,       :decimal
+      t.column :coord_radio,   :decimal
       t.column :heatmap_id,    :integer
       t.timestamps
     end
